requirements.txt
openai
python-dotenv
pandas
tabulate